rename: JpaConfig config 패키지로 이동
